refactor: dedicated test files
