+++ core/src/lib_tests.rs
@@ -1,0 +1,6 @@
+//use crate::{Message, get_message};
+
+#[test]
+fn placeholder_test() {
+    assert_eq!(1, 1);
+}
