+++ core/src/project/reference_tests.rs
@@ -1,0 +1,6 @@
+use crate::project::{local_kpar::LocalKParProject, reference::ProjectReference};
+#[test]
+fn test_kpar() {
+    let kpar = ProjectReference::new(LocalKParProject::new("path", "root").unwrap());
+    let _clone = kpar.clone();
+}
